+++ src/com/devedu/springdemo/dao/CustomerDAO.java
@@ -10,6 +10,6 @@ public interface CustomerDAO {
 
 	public void saveCustomer(Customer customer);
 
-	public Customer getCustomer(int id);
+	public Customer getCustomer(Long id);
 
 }
