+++ src/com/devedu/springdemo/service/CustomerService.java
@@ -10,6 +10,6 @@ public interface CustomerService {
 
 	public void saveCustomer(Customer customer);
 
-	public Customer getCustomer(int id);
+	public Customer getCustomer(Long id);
 
 }
